Separate instruction set from VM
